requirements.txt
opencv-python
pytesseract
numpy
qrcode
pillow